fix: highlight property does not automatically add interactability
If the GameObject has not been an interactable object before, when adding the HighlightProperty
it adds the InteractableObject component but disables the IsTouchable, IsGrabbable,
and IsUsable flags. Besides, the added rigidbody is set to kinematic.
If the GameObject has been an interactable object before, the InteractableObject and rigidbody
stay as they are.
Furthermore, the TouchableProperty, GrabbableProperty, and UsableProperty also
make sure to set the flags properly after being added to the GameObject.